parse libnss for a subid module
If specified use that ofr subid allocations.  By default, use
'files' which is the traditional /etc/subuid and /etc/subgid
allocations.

When non-files nss module is used, functions to edit the range
are not supported.  It may make sense to support it, but it also
may make sense to require another tool to be used.

Have libsubordinateio use nss_ helpers.  This is how for instance
lxc could easily be converted.

Add a test with a dummy module.